TW-344: Update UI for StickyTimestampWidget
